--- datasets/datasets_doc.md
@@ -1,11 +1,0 @@
-
-
-使用したデータセットは，
-
-
-
-Maestro
-> Curtis Hawthorne, Andriy Stasyuk, Adam Roberts, Ian Simon, Cheng-Zhi Anna Huang,<br>
-   Sander Dieleman, Erich Elsen, Jesse Engel, and Douglas Eck. "Enabling<br>
-   Factorized Piano Music Modeling and Generation with the MAESTRO Dataset."<br>
-  In International Conference on Learning Representations, 2019.<br>
