--- hum2midi/hum2midi_doc.md
@@ -1,1 +1,0 @@
-iitaiの管轄
