--- midi_generator/.gitignore
@@ -1,8 +1,0 @@
-/model_maker/models/*
-/model_maker/midi_chunk/*
-/model_maker/tokenizers/*
-/model_maker/generated/*
-/model_maker/prompt/*
-*.mid
-
-datasets_path.json
