+++ model_maker/.gitignore
@@ -1,0 +1,8 @@
+/models/*
+/midi_chunk/*
+/tokenizers/*
+/generated/*
+/prompt/*
+/dataset/*
+
+datasets_path.json
