+++ src/hum2midi/hum2midi_doc.md
@@ -1,0 +1,1 @@
+iitaiの管轄
